requirements.txt
backports.ssl-match-hostname==3.5.0.1
bleach==3.3.1
cached-property==1.3.0
cffi==1.12.3
colorama==0.3.7
configparser==3.8.1
Django==1.10
django-bootstrap3==7.0.1
django-ckeditor==5.3.0
django-crontab==0.7.1
django-datetime-widget==0.9.3
django-js-asset==1.2.2
django-mailer==1.2.2
django-recaptcha2==1.4.1
docker==2.2.0
docker-compose==1.12.0rc1
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
enum34==1.1.6
flake8==3.7.8
funcsigs==1.0.2
functools32==3.2.3.post2
gunicorn==19.7.1
idna==2.8
ipaddress==1.0.18
jsonschema==2.6.0
lockfile==0.12.2
mccabe==0.6.1
mock==2.0.0
packaging==20.9
pbr==5.4.2
pyasn1==0.4.7
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2016.6.1
PyYAML==3.12
requests==2.11.1
six==1.10.0
texttable==0.8.7
typing==3.7.4.1
webencodings==0.5.1
websocket-client==0.40.0
whitenoise==3.3.0
